Fixing semantics of lzss_copy and call sites
`distance` and `length` parameters are in `stream_t` units and relative to current position on stream. Previous semantics of `offset` were unclear and incorrectly applied for Comper/ComperX because it was mixing an offset in `stream_t` units with a position (in bytes) on the stream.

Also added RLE optimization for lzss_copy with distance 1.