Introduce unsafe_forge_span for safe C-style array handling and update related files
This is a workaround for the single dumbest most useless warning in the world. This is literally the whole point of this constructor: to bundle a C-style array plus size into a std::span for safe usage.

They might as well reword the warning to say "This constructor is doing exactly what you want, but we are going to warn about it anyway. And make you do dumb workarounds to avoid the warning."